Updated WeightedSparse, how conservation corrections work, etc.
Now... A WeightedSparse matrix is marked as CONSERVATIVE (or not).  WeightedSparse::apply_e() takes a force_conservation argument.  If force_conservation is in place, then conservation will be forced, using WeightedSparse::wM/Mw; which are not necessarily the same as the row- and column-sum of M.  (They will equal the row and column sum in the case of conservation corrections due to smoothing.  But when correcting due to inconsistent ice sheets, they will be based on the respective ice sheet to which they pertain.)