Fixed bug that used dimEOp in place of dimEOm.
dimEOp is a superset of dimEOm; therefore it was formerly considered "harmless" to use the larger in place of the smaller.  This turned out NOT to be the case, as inclusion in dimEOm affects scaling.  This change properly sets up dimEOm and dimAOm as separate from dimEOp and dimAOp; and properly converts matrices to this new space when needed.  It seems to solve the observed problem of FHC not adding to 1 in the TOPOA file.