Revert changes, to go back to using (jm,im) instead of (lat,lon).
Changing to (lat,lon) everywhere is too pervasive; and inspection of ModeLE input files shows no consistency on this front anyway.  It is closer to CF conventions to use (lat,lon) if you have dimension variables named lat and lon.  But that is a small issue.

This reverts (small) portions of commit 596d822bed881484767b5f34d996009fb6edae66.